Normalize text in Key Generation Procedure sections
It's not clear why the ML-* key generation sentences were sigificantly
different from the ECC key generation sentences.

It was also not clear to me what "the relative specification" meant,
so i've dropped that term.

The sections referenced don't seem to actually "specify formats",
rather they describe the sizes of the objects, so i've adjusted that
text too.