Clarify that component signatures are not valid *OpenPGP* signatures
The current text makes a bold (and i think unsupportable) claim that
each individual component is not a valid signature at all for any
message.

But the argument around it (and, as i understand the situation) is
that the signature can't be a valid OpenPGP signature any longer.  I
could cook up some other signature scheme for which this component
signature *is* a valid signature, but we don't try to address that.